--- discovery-service/src/main/java/org/zowe/apiml/discovery/config/HttpWebSecurityConfig.java
@@ -1,207 +1,0 @@
-/*
- * This program and the accompanying materials are made available under the terms of the
- * Eclipse Public License v2.0 which accompanies this distribution, and is available at
- * https://www.eclipse.org/legal/epl-v20.html
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Copyright Contributors to the Zowe Project.
- */
-
-package org.zowe.apiml.discovery.config;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import org.springframework.context.support.MessageSourceAccessor;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.SpringSecurityMessageSource;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
-import org.zowe.apiml.security.common.config.HandlerInitializer;
-import org.zowe.apiml.security.common.content.BasicContentFilter;
-
-import java.nio.ByteBuffer;
-import java.nio.CharBuffer;
-import java.nio.charset.StandardCharsets;
-import java.security.MessageDigest;
-import java.util.Collections;
-
-/**
- * Main class configuring Spring security for Discovery Service
- * <p>
- * This configuration is applied if "https" Spring profile is not active
- */
-@Slf4j
-@Configuration
-@ComponentScan({
-    "org.zowe.apiml.security.common",
-    "org.zowe.apiml.gateway.security.login"
-})
-@EnableWebSecurity
-@RequiredArgsConstructor
-@Profile({"!https", "!attlsServer", "!attlsClient"})
-public class HttpWebSecurityConfig extends AbstractWebSecurityConfigurer {
-    private static final String DISCOVERY_REALM = "API Mediation Discovery Service realm";
-
-    @Value("${apiml.discovery.userid:#{null}}")
-    private String discoveryUserId;
-
-    @Value("${apiml.discovery.password:#{null}}")
-    private char[] discoveryPassword;
-
-    @Value("${apiml.metrics.enabled:false}")
-    private boolean isMetricsEnabled;
-
-    @Value("${apiml.health.protected:false}")
-    private boolean isHealthEndpointProtected;
-
-    @Autowired
-    public void configureGlobal(AuthenticationManagerBuilder auth) {
-        // we cannot use `auth.inMemoryAuthentication()` because it does not support char array
-        auth.authenticationProvider(new EurekaBasicAuthenticationProvider(discoveryUserId, discoveryPassword));
-    }
-
-    private final HandlerInitializer handlerInitializer;
-
-    @Bean
-    public WebSecurityCustomizer httpWebSecurityCustomizer() {
-        AntPathRequestMatcher[] noSecurityAntMatchers = {
-            new AntPathRequestMatcher("/favicon.ico"),
-            new AntPathRequestMatcher("/eureka/css/**"),
-            new AntPathRequestMatcher("/eureka/js/**"),
-            new AntPathRequestMatcher("/eureka/fonts/**"),
-            new AntPathRequestMatcher("/eureka/images/**")
-        };
-        return web -> web.ignoring().requestMatchers(noSecurityAntMatchers);
-    }
-
-    @Bean
-    public SecurityFilterChain httpFilterChain(HttpSecurity http) throws Exception {
-
-        if (!isHealthEndpointProtected) {
-            http.authorizeHttpRequests(requests -> requests
-                .requestMatchers(new AntPathRequestMatcher("/application/health")).permitAll());
-        }
-
-        baseConfigure(http)
-            .httpBasic(basic -> basic.realmName(DISCOVERY_REALM))
-            .authorizeHttpRequests(requests -> requests
-                .requestMatchers(new AntPathRequestMatcher("/application/info")).permitAll()
-                .requestMatchers(new AntPathRequestMatcher("/**")).authenticated());
-
-        if (isMetricsEnabled) {
-            http.authorizeHttpRequests(requests -> requests.requestMatchers(new AntPathRequestMatcher("/application/hystrixstream")).permitAll());
-        }
-
-        return http.apply(new CustomSecurityFilters()).and().build();
-    }
-
-    private class CustomSecurityFilters extends AbstractHttpConfigurer<CustomSecurityFilters, HttpSecurity> {
-        @Override
-        public void configure(HttpSecurity http) {
-            AuthenticationManager authenticationManager = http.getSharedObject(AuthenticationManager.class);
-            http.addFilterBefore(basicFilter(authenticationManager), UsernamePasswordAuthenticationFilter.class);
-        }
-
-        private BasicContentFilter basicFilter(AuthenticationManager authenticationManager) {
-            return new BasicContentFilter(authenticationManager, handlerInitializer.getAuthenticationFailureHandler(), handlerInitializer.getResourceAccessExceptionHandler());
-        }
-    }
-
-    static class EurekaBasicAuthenticationProvider implements AuthenticationProvider {
-
-        private final boolean discoveryCredentialsProvider;
-        private final byte[] discoveryUserid;
-        private final byte[] discoveryPassword;
-
-        private final MessageSourceAccessor messages = SpringSecurityMessageSource.getAccessor();
-
-        EurekaBasicAuthenticationProvider(String discoveryUserid, char[] discoveryPassword) {
-            this.discoveryUserid = getBytes(discoveryUserid);
-            this.discoveryPassword = getBytes(discoveryPassword);
-
-            discoveryCredentialsProvider = (this.discoveryUserid.length > 0) && (this.discoveryPassword.length > 0);
-            if (!discoveryCredentialsProvider) {
-                log.warn("Eureka credentials are not set. Please configure properties `apiml.discovery.userid` and `apiml.discovery.password` or change type of Eureka authentication.");
-            }
-        }
-
-        private boolean isValid(Authentication authentication) {
-            byte[] userId = getUser(authentication);
-            boolean userMatching = MessageDigest.isEqual(userId, discoveryUserid);
-
-            byte[] password = getPassword(authentication);
-            boolean passwordMatching = MessageDigest.isEqual(password, discoveryPassword);
-
-            return discoveryCredentialsProvider && userMatching && passwordMatching;
-        }
-
-        static byte[] getBytes(Object obj) {
-            if (obj == null) {
-                return new byte[0];
-            }
-            if (obj instanceof byte[]) {
-                return (byte[]) obj;
-            }
-            if (obj instanceof char[]) {
-                ByteBuffer byteBuffer = StandardCharsets.UTF_8.encode(CharBuffer.wrap((char[]) obj));
-                byte[] bytes = new byte[byteBuffer.limit()];
-                byteBuffer.get(bytes);
-                return bytes;
-            }
-            if (obj instanceof String) {
-                return ((String) obj).getBytes(StandardCharsets.UTF_8);
-            }
-            return String.valueOf(obj).getBytes(StandardCharsets.UTF_8);
-        }
-
-        private byte[] getPassword(Authentication authentication) {
-            return getBytes(authentication.getCredentials());
-        }
-
-        private byte[] getUser(Authentication authentication) {
-            return getBytes(authentication.getPrincipal());
-        }
-
-        @Override
-        public Authentication authenticate(Authentication authentication) throws AuthenticationException {
-            if (isValid(authentication)) {
-                UsernamePasswordAuthenticationToken result = UsernamePasswordAuthenticationToken.authenticated(
-                    authentication.getPrincipal(),
-                    authentication.getCredentials(),
-                    Collections.singleton(new SimpleGrantedAuthority("EUREKA"))
-                );
-                result.setDetails(authentication.getDetails());
-                return result;
-            }
-
-            throw new BadCredentialsException(this.messages
-                .getMessage("AbstractUserDetailsAuthenticationProvider.badCredentials", "Bad credentials"));
-        }
-
-        @Override
-        public boolean supports(Class<?> authentication) {
-            return (UsernamePasswordAuthenticationToken.class.isAssignableFrom(authentication));
-        }
-
-    }
-
-}
