+++ discovery-service/src/main/java/org/zowe/apiml/discovery/config/HttpsWebSecurityConfig.java
@@ -11,16 +11,24 @@
 package org.zowe.apiml.discovery.config;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
+import org.springframework.context.support.MessageSourceAccessor;
 import org.springframework.core.annotation.Order;
 import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.core.SpringSecurityMessageSource;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.web.SecurityFilterChain;
@@ -38,6 +46,10 @@ import org.zowe.apiml.security.common.content.BasicContentFilter;
 import org.zowe.apiml.security.common.content.BearerContentFilter;
 import org.zowe.apiml.security.common.content.CookieContentFilter;
 
+import java.nio.ByteBuffer;
+import java.nio.CharBuffer;
+import java.nio.charset.StandardCharsets;
+import java.security.MessageDigest;
 import java.util.Collections;
 
 /**
@@ -49,7 +61,6 @@ import java.util.Collections;
 @RequiredArgsConstructor
 @EnableWebSecurity
 @EnableApimlAuth
-@Profile({"https", "attlsServer"})
 public class HttpsWebSecurityConfig extends AbstractWebSecurityConfigurer {
 
     private final HandlerInitializer handlerInitializer;
@@ -138,7 +149,7 @@ public class HttpsWebSecurityConfig extends AbstractWebSecurityConfigurer {
                 http.addFilterBefore(new SecureConnectionFilter(), AttlsFilter.class);
             }
         } else {
-            http.authenticationProvider(new HttpWebSecurityConfig.EurekaBasicAuthenticationProvider(discoveryUserId, discoveryPassword))
+            http.authenticationProvider(new EurekaBasicAuthenticationProvider(discoveryUserId, discoveryPassword))
                 .authorizeHttpRequests(requests -> requests.anyRequest().authenticated())
                 .httpBasic(basic -> basic.realmName(DISCOVERY_REALM));
 
@@ -210,4 +221,84 @@ public class HttpsWebSecurityConfig extends AbstractWebSecurityConfigurer {
     private UserDetailsService x509UserDetailsService() {
         return username -> new User("eurekaClient", "", Collections.emptyList());
     }
+
+    @Slf4j
+    static class EurekaBasicAuthenticationProvider implements AuthenticationProvider {
+
+        private final boolean discoveryCredentialsProvider;
+        private final byte[] discoveryUserid;
+        private final byte[] discoveryPassword;
+
+        private final MessageSourceAccessor messages = SpringSecurityMessageSource.getAccessor();
+
+        EurekaBasicAuthenticationProvider(String discoveryUserid, char[] discoveryPassword) {
+            this.discoveryUserid = getBytes(discoveryUserid);
+            this.discoveryPassword = getBytes(discoveryPassword);
+
+            discoveryCredentialsProvider = (this.discoveryUserid.length > 0) && (this.discoveryPassword.length > 0);
+            if (!discoveryCredentialsProvider) {
+                log.warn("Eureka credentials are not set. Please configure properties `apiml.discovery.userid` and `apiml.discovery.password` or change type of Eureka authentication.");
+            }
+        }
+
+        private boolean isValid(Authentication authentication) {
+            byte[] userId = getUser(authentication);
+            boolean userMatching = MessageDigest.isEqual(userId, discoveryUserid);
+
+            byte[] password = getPassword(authentication);
+            boolean passwordMatching = MessageDigest.isEqual(password, discoveryPassword);
+
+            return discoveryCredentialsProvider && userMatching && passwordMatching;
+        }
+
+        static byte[] getBytes(Object obj) {
+            if (obj == null) {
+                return new byte[0];
+            }
+            if (obj instanceof byte[]) {
+                return (byte[]) obj;
+            }
+            if (obj instanceof char[]) {
+                ByteBuffer byteBuffer = StandardCharsets.UTF_8.encode(CharBuffer.wrap((char[]) obj));
+                byte[] bytes = new byte[byteBuffer.limit()];
+                byteBuffer.get(bytes);
+                return bytes;
+            }
+            if (obj instanceof String) {
+                return ((String) obj).getBytes(StandardCharsets.UTF_8);
+            }
+            return String.valueOf(obj).getBytes(StandardCharsets.UTF_8);
+        }
+
+        private byte[] getPassword(Authentication authentication) {
+            return getBytes(authentication.getCredentials());
+        }
+
+        private byte[] getUser(Authentication authentication) {
+            return getBytes(authentication.getPrincipal());
+        }
+
+        @Override
+        public Authentication authenticate(Authentication authentication) throws AuthenticationException {
+            if (isValid(authentication)) {
+                UsernamePasswordAuthenticationToken result = UsernamePasswordAuthenticationToken.authenticated(
+                    authentication.getPrincipal(),
+                    authentication.getCredentials(),
+                    Collections.singleton(new SimpleGrantedAuthority("EUREKA"))
+                );
+                result.setDetails(authentication.getDetails());
+                return result;
+            }
+
+            throw new BadCredentialsException(this.messages
+                .getMessage("AbstractUserDetailsAuthenticationProvider.badCredentials", "Bad credentials"));
+        }
+
+        @Override
+        public boolean supports(Class<?> authentication) {
+            return (UsernamePasswordAuthenticationToken.class.isAssignableFrom(authentication));
+        }
+
+    }
+
 }
